requirements.txt
requests==2.32.3
python-dotenv==1.0.1
openai==1.35.13
